Add subdirectories in views directory
