+++ lib/views/modal/progress-modal-view.js
@@ -4,7 +4,7 @@
 
 import { File } from 'atom'
 import etch from 'etch'
-import ViewBase from './view-base'
+import ViewBase from '../view-base'
 
 export default class ProgressModalView extends ViewBase {
   render() {
