+++ lib/views/section/docbase-section-view.js
@@ -7,33 +7,35 @@ import etch from 'etch'
 import fs from 'fs-plus'
 import stringify from 'json-stable-stringify'
 import SectionViewBase from './section-view-base'
-import QiitaService from '../services/qiita-service'
-import ConfigUtil from '../utils/config-util'
+import DocBaseService from '../../services/docbase-service'
+import ConfigUtil from '../../utils/config-util'
 
-export default class QiitaSectionView extends SectionViewBase {
+const docbaseScopes = ['everyone', 'group', 'private']
+
+export default class DocBaseSectionView extends SectionViewBase {
   render() {
     return (
-      <section className="qiita service">
-        <h2>Sync with Qiita</h2>
+      <section className="docbase service">
+        <h2>Sync with DocBase</h2>
 
-        <label htmlFor="qiita-access-token">
+        <label htmlFor="docbase-access-token">
           <span>AccessToken</span>
           <input
             type="text"
-            id="qiita-access-token"
-            ref="qiitaAccessToken"
-            value={this.props.qiitaAccessToken || ''}
+            id="docbase-access-token"
+            ref="docbaseAccessToken"
+            value={this.props.docbaseAccessToken || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="qiita-user-name">
-          <span>UserName</span>
+        <label htmlFor="docbase-domain">
+          <span>Domain</span>
           <input
             type="text"
-            id="qiita-user-name"
-            ref="qiitaUserName"
-            value={this.props.qiitaUserName || ''}
+            id="docbase-domain"
+            ref="docbaseDomain"
+            value={this.props.docbaseDomain || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
@@ -41,71 +43,105 @@ export default class QiitaSectionView extends SectionViewBase {
         <h3>
           Item properties&nbsp;
           <a
-            href={this.props.qiitaItemUrl || ''}
-            style={{ visibility: (this.props.qiitaItemUrl === null) ? 'hidden' : 'visible' }}
+            href={this.props.docbaseItemUrl || ''}
+            style={{ visibility: (this.props.docbaseItemUrl === null) ? 'hidden' : 'visible' }}
             tabIndex={this.props.startTabIndex++}
           >
             <span role="img" aria-label="Link">🔗</span>
           </a>
         </h3>
 
-        <label htmlFor="qiita-title">
+        <label htmlFor="docbase-title">
           <span>Title</span>
           <input
             type="text"
-            id="qiita-title"
-            ref="qiitaTitle"
-            value={this.props.qiitaTitle || ''}
+            id="docbase-title"
+            ref="docbaseTitle"
+            value={this.props.docbaseTitle || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="qiita-tags">
+        <label htmlFor="docbase-tags">
           <span>Tags</span>
           <input
             type="text"
-            id="qiita-tags"
-            ref="qiitaTags"
-            value={this.props.qiitaTags || ''}
+            id="docbase-tags"
+            ref="docbaseTags"
+            value={this.props.docbaseTags || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <button id="qiita-import" on={{ click: this.importFromQiita }} tabIndex={this.props.startTabIndex++}>
+        <label htmlFor="docbase-groups">
+          <span>Groups</span>
+          <input
+            type="text"
+            id="docbase-groups"
+            ref="docbaseGroups"
+            value={this.props.docbaseGroups || ''}
+            tabIndex={this.props.startTabIndex++}
+          />
+        </label><br />
+
+        <label htmlFor="docbase-scope">
+          <span>Scope</span>
+          <select
+            id="docbase-scope"
+            ref="docbaseScope"
+            value={this.props.docbaseScope || ''}
+            tabIndex={this.props.startTabIndex++}
+          >
+            {
+              docbaseScopes.map((scope) => {
+                if (scope === this.props.docbaseScope) {
+                  return <option value={scope} selected>{scope}</option>
+                }
+                return <option value={scope}>{scope}</option>
+              })
+            }
+          </select>
+        </label><br />
+
+        <button id="docbase-import" on={{ click: this.importFromDocBase }} tabIndex={this.props.startTabIndex++}>
           Import
         </button>
-        <button id="qiita-export" on={{ click: this.exportToQiita }} tabIndex={this.props.startTabIndex++}>
+        <button id="docbase-export" on={{ click: this.exportToDocBase }} tabIndex={this.props.startTabIndex++}>
           Export
         </button>
       </section>
     )
   }
 
-  importFromQiita(event) {
-    this.handleQiitaEvent(event, 'import')
+  importFromDocBase(event) {
+    this.handleDocBaseEvent(event, 'import')
   }
 
-  exportToQiita(event) {
-    this.handleQiitaEvent(event, 'export')
+  exportToDocBase(event) {
+    this.handleDocBaseEvent(event, 'export')
   }
 
-  handleQiitaEvent(event, type = 'import') {
+  handleDocBaseEvent(event, type = 'import') {
     try {
-      console.log(`Start Qiita ${type}.`)
+      console.log(`Start DocBase ${type}.`)
       const { localPath, configDirPath, optionKeepFilePath } = this.props
-      const accessToken = this.refs.qiitaAccessToken.value
-      const userName = this.refs.qiitaUserName.value
-      const qiitaTitle = this.refs.qiitaTitle.value
-      const qiitaTags = this.refs.qiitaTags.value
-      QiitaSectionView.startHandlingQiitaEvent(
+      const accessToken = this.refs.docbaseAccessToken.value
+      const domain = this.refs.docbaseDomain.value
+      const docbaseTitle = this.refs.docbaseTitle.value
+      const docbaseTags = this.refs.docbaseTags.value
+      const docbaseGroups = this.refs.docbaseGroups.value
+      const docbaseScope = this.refs.docbaseScope.value
+      DocBaseSectionView.startHandlingDocBaseEvent(
         type,
         configDirPath,
         localPath,
         optionKeepFilePath,
         accessToken,
-        userName,
-        qiitaTitle,
-        qiitaTags,
+        domain,
+        docbaseTitle,
+        docbaseTags,
+        docbaseGroups,
+        docbaseScope,
       )
     } catch (error) {
       atom.notifications.addError('Something went wrong.')
@@ -113,23 +149,25 @@ export default class QiitaSectionView extends SectionViewBase {
     }
   }
 
-  static startHandlingQiitaEvent(
+  static startHandlingDocBaseEvent(
     type,
     configDirPath,
     localPath,
     optionKeepFilePath,
     accessToken,
-    userName,
-    qiitaTitle,
-    qiitaTags,
+    domain,
+    docbaseTitle,
+    docbaseTags,
+    docbaseGroups,
+    docbaseScope,
   ) {
     if (localPath === '') {
       atom.notifications.addError('LocalPath must be specified.')
       return
     }
 
-    if (type === 'export' && qiitaTags === '') {
-      atom.notifications.addError('Need to specify tag in exporting.')
+    if (type === 'export' && !docbaseScopes.includes(docbaseScope)) {
+      atom.notifications.addError(`Invalid scope specified: ${docbaseScope}`)
       return
     }
 
@@ -152,28 +190,28 @@ export default class QiitaSectionView extends SectionViewBase {
 
     const loadConfigPromise = configFile.read().then((configBody) => {
       const config = JSON.parse(configBody)
-      config.qiita = config.qiita || {}
-      config.qiita.userName = userName
+      config.docbase = config.docbase || {}
+      config.docbase.domain = domain
       // TODO For legacy implementation. To be removed.
-      delete config.qiita.accessToken
+      delete config.docbase.accessToken
       return config
     })
     const loadSecretConfigPromise = secretConfigFile.read().then((configBody) => {
       const config = JSON.parse(configBody)
-      config.qiita = config.qiita || {}
+      config.docbase = config.docbase || {}
       if (atom.config.get('sync-sync.keepSecrets')) {
-        config.qiita.accessToken = accessToken
+        config.docbase.accessToken = accessToken
       } else {
-        delete config.qiita.accessToken
+        delete config.docbase.accessToken
       }
       return config
     })
 
     Promise.all([loadConfigPromise, loadSecretConfigPromise]).then((configs) => {
       const [config, secretConfig] = configs
-      const itemsConfig = config.qiita.items || {}
+      const itemsConfig = config.docbase.items || {}
 
-      const qiitaService = new QiitaService(accessToken)
+      const docbaseService = new DocBaseService(accessToken)
       const progressModal = this.showProgressModal('Now Processing', 'Please wait for the process finished...')
       let handledProcess
 
@@ -194,7 +232,7 @@ export default class QiitaSectionView extends SectionViewBase {
         //
 
         // Call listing API.
-        handledProcess = qiitaService.listAllItems(userName)
+        handledProcess = docbaseService.listAllItems(domain)
           .then((items) => {
             // Save files.
             const fileSaveProcesses = []
@@ -219,10 +257,12 @@ export default class QiitaSectionView extends SectionViewBase {
               itemsConfig[item.id] = {
                 path: relativeItemPath,
                 url: item.url,
-                updatedAt: item.updated_at,
+                createdAt: item.created_at,
                 title: item.title,
                 tags: item.tags.map(tag => tag.name),
-                userName,
+                groups: item.groups,
+                scope: item.scope,
+                domain,
               }
 
               const fileSaveProcess = file.write(this.normalizeNewLine(item.body)).then(() => {
@@ -252,7 +292,7 @@ export default class QiitaSectionView extends SectionViewBase {
           }
 
           // Call get API.
-          handledProcess = qiitaService.getItem(itemId)
+          handledProcess = docbaseService.getItem(domain, itemId)
             .then((item) => {
               if (targetFile.existsSync()) {
                 // Check not-synced local modification.
@@ -269,9 +309,11 @@ export default class QiitaSectionView extends SectionViewBase {
               // Save file.
               return targetFile.write(this.normalizeNewLine(item.body)).then(() => {
                 console.log('File saved: ', targetFile.getPath())
-                itemsConfig[item.id].updatedAt = item.updated_at
+                itemsConfig[item.id].createdAt = item.created_at
                 itemsConfig[item.id].title = item.title
                 itemsConfig[item.id].tags = item.tags.map(tag => tag.name)
+                itemsConfig[item.id].groups = item.groups
+                itemsConfig[item.id].scope = item.scope
                 itemsConfig[item.id].digest = targetFile.getDigestSync()
                 return Promise.resolve()
               })
@@ -282,26 +324,29 @@ export default class QiitaSectionView extends SectionViewBase {
           //
           handledProcess = targetFile.read().then((fileBody) => {
             const relativeItemPath = configDir.relativize(localPath)
-            let title = qiitaTitle
+            let title = docbaseTitle
             if (!title || title === '') {
               title = targetFile.getBaseName().replace(/\.[^/.]+$/, '') // remove extension
             }
             const itemId = Object.keys(itemsConfig)
               .find(key => itemsConfig[key].path === relativeItemPath)
-            const tags = qiitaTags.split(',').map(tag => tag.trim())
+            const tags = docbaseTags.split(',').map(tag => tag.trim())
+            const groups = docbaseGroups.split(',').map(group => group.trim())
             if (!itemId) {
               // In case of posting.
               // Call post API.
-              return qiitaService.postItem(title, fileBody, tags)
+              return docbaseService.postItem(domain, title, fileBody, tags, groups, docbaseScope)
                 .then((item) => {
                   // Save meta data to config.
                   itemsConfig[item.id] = {
                     path: relativeItemPath,
                     url: item.url,
-                    updatedAt: item.updated_at,
+                    createdAt: item.created_at,
                     title: item.title,
                     tags: item.tags.map(tag => tag.name),
-                    userName,
+                    groups: item.groups,
+                    scope: item.scope,
+                    domain,
                     digest: targetFile.getDigestSync(),
                   }
                   return Promise.resolve()
@@ -310,12 +355,14 @@ export default class QiitaSectionView extends SectionViewBase {
 
             // In case of updating.
             // Call update API.
-            return qiitaService.updateItem(itemId, title, fileBody, tags)
+            return docbaseService.updateItem(domain, itemId, title, fileBody, tags, groups, docbaseScope)
               .then((item) => {
                 // Save meta data to config.
-                itemsConfig[item.id].updatedAt = item.updated_at
+                itemsConfig[item.id].createdAt = item.created_at
                 itemsConfig[item.id].title = item.title
                 itemsConfig[item.id].tags = item.tags.map(tag => tag.name)
+                itemsConfig[item.id].groups = item.groups
+                itemsConfig[item.id].scope = item.scope
                 itemsConfig[item.id].digest = targetFile.getDigestSync()
                 return Promise.resolve()
               })
@@ -325,7 +372,7 @@ export default class QiitaSectionView extends SectionViewBase {
 
       handledProcess.then(() => {
         // Save meta data to config.
-        config.qiita.items = itemsConfig
+        config.docbase.items = itemsConfig
         // See [javascript - sort object properties and JSON.stringify - Stack Overflow](https://stackoverflow.com/questions/16167581/).
         const configSavePromise = configFile.write(stringify(config, { space: 2 }))
         const secretConfigSavePromise = secretConfigFile.write(stringify(secretConfig, { space: 2 }))
@@ -337,7 +384,7 @@ export default class QiitaSectionView extends SectionViewBase {
       }).catch((error) => {
         // Handle error.
         atom.notifications.addError('Something went wrong.')
-        console.error('Error occurred in QiitaService: ', error)
+        console.error('Error occurred in DocBaseService: ', error)
         progressModal.destroy()
       })
     })
