+++ lib/views/section/esa-section-view.js
@@ -7,35 +7,33 @@ import etch from 'etch'
 import fs from 'fs-plus'
 import stringify from 'json-stable-stringify'
 import SectionViewBase from './section-view-base'
-import DocBaseService from '../services/docbase-service'
-import ConfigUtil from '../utils/config-util'
+import EsaService from '../../services/esa-service'
+import ConfigUtil from '../../utils/config-util'
 
-const docbaseScopes = ['everyone', 'group', 'private']
-
-export default class DocBaseSectionView extends SectionViewBase {
+export default class EsaSectionView extends SectionViewBase {
   render() {
     return (
-      <section className="docbase service">
-        <h2>Sync with DocBase</h2>
+      <section className="esa service">
+        <h2>Sync with esa.io</h2>
 
-        <label htmlFor="docbase-access-token">
+        <label htmlFor="esa-access-token">
           <span>AccessToken</span>
           <input
             type="text"
-            id="docbase-access-token"
-            ref="docbaseAccessToken"
-            value={this.props.docbaseAccessToken || ''}
+            id="esa-access-token"
+            ref="esaAccessToken"
+            value={this.props.esaAccessToken || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="docbase-domain">
-          <span>Domain</span>
+        <label htmlFor="esa-team-name">
+          <span>TeamName</span>
           <input
             type="text"
-            id="docbase-domain"
-            ref="docbaseDomain"
-            value={this.props.docbaseDomain || ''}
+            id="esa-team-name"
+            ref="esaTeamName"
+            value={this.props.esaTeamName || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
@@ -43,105 +41,84 @@ export default class DocBaseSectionView extends SectionViewBase {
         <h3>
           Item properties&nbsp;
           <a
-            href={this.props.docbaseItemUrl || ''}
-            style={{ visibility: (this.props.docbaseItemUrl === null) ? 'hidden' : 'visible' }}
+            href={this.props.esaItemUrl || ''}
+            style={{ visibility: (this.props.esaItemUrl === null) ? 'hidden' : 'visible' }}
             tabIndex={this.props.startTabIndex++}
           >
             <span role="img" aria-label="Link">🔗</span>
           </a>
         </h3>
 
-        <label htmlFor="docbase-title">
+        <label htmlFor="esa-title">
           <span>Title</span>
           <input
             type="text"
-            id="docbase-title"
-            ref="docbaseTitle"
-            value={this.props.docbaseTitle || ''}
+            id="esa-title"
+            ref="esaTitle"
+            value={this.props.esaTitle || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="docbase-tags">
+        <label htmlFor="esa-tags">
           <span>Tags</span>
           <input
             type="text"
-            id="docbase-tags"
-            ref="docbaseTags"
-            value={this.props.docbaseTags || ''}
+            id="esa-tags"
+            ref="esaTags"
+            value={this.props.esaTags || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="docbase-groups">
-          <span>Groups</span>
+        <label htmlFor="esa-category">
+          <span>Category</span>
           <input
             type="text"
-            id="docbase-groups"
-            ref="docbaseGroups"
-            value={this.props.docbaseGroups || ''}
+            id="esa-category"
+            ref="esaCategory"
+            value={this.props.esaCategory || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="docbase-scope">
-          <span>Scope</span>
-          <select
-            id="docbase-scope"
-            ref="docbaseScope"
-            value={this.props.docbaseScope || ''}
-            tabIndex={this.props.startTabIndex++}
-          >
-            {
-              docbaseScopes.map((scope) => {
-                if (scope === this.props.docbaseScope) {
-                  return <option value={scope} selected>{scope}</option>
-                }
-                return <option value={scope}>{scope}</option>
-              })
-            }
-          </select>
-        </label><br />
-
-        <button id="docbase-import" on={{ click: this.importFromDocBase }} tabIndex={this.props.startTabIndex++}>
+        <button id="esa-import" on={{ click: this.importFromEsa }} tabIndex={this.props.startTabIndex++}>
           Import
         </button>
-        <button id="docbase-export" on={{ click: this.exportToDocBase }} tabIndex={this.props.startTabIndex++}>
+        <button id="esa-export" on={{ click: this.exportToEsa }} tabIndex={this.props.startTabIndex++}>
           Export
         </button>
       </section>
     )
   }
 
-  importFromDocBase(event) {
-    this.handleDocBaseEvent(event, 'import')
+  importFromEsa(event) {
+    this.handleEsaEvent(event, 'import')
   }
 
-  exportToDocBase(event) {
-    this.handleDocBaseEvent(event, 'export')
+  exportToEsa(event) {
+    this.handleEsaEvent(event, 'export')
   }
 
-  handleDocBaseEvent(event, type = 'import') {
+  handleEsaEvent(event, type = 'import') {
     try {
-      console.log(`Start DocBase ${type}.`)
+      console.log(`Start Esa ${type}.`)
       const { localPath, configDirPath, optionKeepFilePath } = this.props
-      const accessToken = this.refs.docbaseAccessToken.value
-      const domain = this.refs.docbaseDomain.value
-      const docbaseTitle = this.refs.docbaseTitle.value
-      const docbaseTags = this.refs.docbaseTags.value
-      const docbaseGroups = this.refs.docbaseGroups.value
-      const docbaseScope = this.refs.docbaseScope.value
-      DocBaseSectionView.startHandlingDocBaseEvent(
+      const accessToken = this.refs.esaAccessToken.value
+      const teamName = this.refs.esaTeamName.value
+      const esaTitle = this.refs.esaTitle.value
+      const esaTags = this.refs.esaTags.value
+      const esaCategory = this.refs.esaCategory.value
+      EsaSectionView.startHandlingEsaEvent(
         type,
         configDirPath,
         localPath,
         optionKeepFilePath,
         accessToken,
-        domain,
-        docbaseTitle,
-        docbaseTags,
-        docbaseGroups,
-        docbaseScope,
+        teamName,
+        esaTitle,
+        esaTags,
+        esaCategory,
       )
     } catch (error) {
       atom.notifications.addError('Something went wrong.')
@@ -149,25 +126,19 @@ export default class DocBaseSectionView extends SectionViewBase {
     }
   }
 
-  static startHandlingDocBaseEvent(
+  static startHandlingEsaEvent(
     type,
     configDirPath,
     localPath,
     optionKeepFilePath,
     accessToken,
-    domain,
-    docbaseTitle,
-    docbaseTags,
-    docbaseGroups,
-    docbaseScope,
+    teamName,
+    esaTitle,
+    esaTags,
+    esaCategory,
   ) {
     if (localPath === '') {
       atom.notifications.addError('LocalPath must be specified.')
-      return
-    }
-
-    if (type === 'export' && !docbaseScopes.includes(docbaseScope)) {
-      atom.notifications.addError(`Invalid scope specified: ${docbaseScope}`)
       return
     }
 
@@ -190,28 +161,28 @@ export default class DocBaseSectionView extends SectionViewBase {
 
     const loadConfigPromise = configFile.read().then((configBody) => {
       const config = JSON.parse(configBody)
-      config.docbase = config.docbase || {}
-      config.docbase.domain = domain
+      config.esa = config.esa || {}
+      config.esa.teamName = teamName
       // TODO For legacy implementation. To be removed.
-      delete config.docbase.accessToken
+      delete config.esa.accessToken
       return config
     })
     const loadSecretConfigPromise = secretConfigFile.read().then((configBody) => {
       const config = JSON.parse(configBody)
-      config.docbase = config.docbase || {}
+      config.esa = config.esa || {}
       if (atom.config.get('sync-sync.keepSecrets')) {
-        config.docbase.accessToken = accessToken
+        config.esa.accessToken = accessToken
       } else {
-        delete config.docbase.accessToken
+        delete config.esa.accessToken
       }
       return config
     })
 
     Promise.all([loadConfigPromise, loadSecretConfigPromise]).then((configs) => {
       const [config, secretConfig] = configs
-      const itemsConfig = config.docbase.items || {}
+      const itemsConfig = config.esa.items || {}
 
-      const docbaseService = new DocBaseService(accessToken)
+      const esaService = new EsaService(accessToken)
       const progressModal = this.showProgressModal('Now Processing', 'Please wait for the process finished...')
       let handledProcess
 
@@ -232,18 +203,23 @@ export default class DocBaseSectionView extends SectionViewBase {
         //
 
         // Call listing API.
-        handledProcess = docbaseService.listAllItems(domain)
+        handledProcess = esaService.listAllItems(teamName)
           .then((items) => {
             // Save files.
             const fileSaveProcesses = []
             items.forEach((item) => {
+              // Use category as directory.
+              let fileDirectory = ''
+              if (item.category && item.category !== '') {
+                fileDirectory = `${item.category}/`
+              }
               let file = new File(`${localPath}/${item.title}.md`)
               if (optionKeepFilePath && itemsConfig[item.id]) {
                 file = new File(`${configDir.getPath()}/${itemsConfig[item.id].path}`)
               }
               if (file.existsSync()) {
                 // Check not-synced local modification.
-                const syncedDigest = (itemsConfig[item.id]) ? itemsConfig[item.id].digest : null
+                const syncedDigest = (itemsConfig[item.number]) ? itemsConfig[item.number].digest : null
                 if (file.getDigestSync() !== syncedDigest) {
                   atom.notifications.addError(
                     'Cannot import because of not-synced local modification.',
@@ -254,20 +230,22 @@ export default class DocBaseSectionView extends SectionViewBase {
               }
 
               const relativeItemPath = configDir.relativize(file.getPath())
-              itemsConfig[item.id] = {
+              itemsConfig[item.number] = {
                 path: relativeItemPath,
                 url: item.url,
-                createdAt: item.created_at,
-                title: item.title,
-                tags: item.tags.map(tag => tag.name),
-                groups: item.groups,
-                scope: item.scope,
-                domain,
+                updatedAt: item.updated_at,
+                title: item.name,
+                tags: item.tags,
+                category: item.category,
+                revision: item.revision_number,
+                wip: item.wip,
+                message: item.message,
+                teamName,
               }
 
-              const fileSaveProcess = file.write(this.normalizeNewLine(item.body)).then(() => {
+              const fileSaveProcess = file.write(this.normalizeNewLine(item.body_md)).then(() => {
                 console.log('File saved: ', file.getPath())
-                itemsConfig[item.id].digest = file.getDigestSync()
+                itemsConfig[item.number].digest = file.getDigestSync()
                 return Promise.resolve()
               })
               fileSaveProcesses.push(fileSaveProcess)
@@ -292,11 +270,11 @@ export default class DocBaseSectionView extends SectionViewBase {
           }
 
           // Call get API.
-          handledProcess = docbaseService.getItem(domain, itemId)
+          handledProcess = esaService.getItem(teamName, itemId)
             .then((item) => {
               if (targetFile.existsSync()) {
                 // Check not-synced local modification.
-                const syncedDigest = (itemsConfig[item.id]) ? itemsConfig[item.id].digest : null
+                const syncedDigest = (itemsConfig[item.number]) ? itemsConfig[item.number].digest : null
                 if (targetFile.getDigestSync() !== syncedDigest) {
                   atom.notifications.addError(
                     'Cannot import because of not-synced local modification.',
@@ -307,14 +285,16 @@ export default class DocBaseSectionView extends SectionViewBase {
               }
 
               // Save file.
-              return targetFile.write(this.normalizeNewLine(item.body)).then(() => {
+              return targetFile.write(this.normalizeNewLine(item.body_md)).then(() => {
                 console.log('File saved: ', targetFile.getPath())
-                itemsConfig[item.id].createdAt = item.created_at
-                itemsConfig[item.id].title = item.title
-                itemsConfig[item.id].tags = item.tags.map(tag => tag.name)
-                itemsConfig[item.id].groups = item.groups
-                itemsConfig[item.id].scope = item.scope
-                itemsConfig[item.id].digest = targetFile.getDigestSync()
+                itemsConfig[item.number].updatedAt = item.updated_at
+                itemsConfig[item.number].title = item.name
+                itemsConfig[item.number].tags = item.tags
+                itemsConfig[item.number].category = item.category
+                itemsConfig[item.number].revision = item.revision_number
+                itemsConfig[item.number].wip = item.wip
+                itemsConfig[item.number].message = item.message
+                itemsConfig[item.number].digest = targetFile.getDigestSync()
                 return Promise.resolve()
               })
             })
@@ -324,29 +304,30 @@ export default class DocBaseSectionView extends SectionViewBase {
           //
           handledProcess = targetFile.read().then((fileBody) => {
             const relativeItemPath = configDir.relativize(localPath)
-            let title = docbaseTitle
+            let title = esaTitle
             if (!title || title === '') {
               title = targetFile.getBaseName().replace(/\.[^/.]+$/, '') // remove extension
             }
             const itemId = Object.keys(itemsConfig)
               .find(key => itemsConfig[key].path === relativeItemPath)
-            const tags = docbaseTags.split(',').map(tag => tag.trim())
-            const groups = docbaseGroups.split(',').map(group => group.trim())
+            const tags = esaTags.split(',').map(tag => tag.trim())
             if (!itemId) {
               // In case of posting.
               // Call post API.
-              return docbaseService.postItem(domain, title, fileBody, tags, groups, docbaseScope)
+              return esaService.postItem(teamName, title, fileBody, tags, esaCategory)
                 .then((item) => {
                   // Save meta data to config.
-                  itemsConfig[item.id] = {
+                  itemsConfig[item.number] = {
                     path: relativeItemPath,
                     url: item.url,
-                    createdAt: item.created_at,
-                    title: item.title,
-                    tags: item.tags.map(tag => tag.name),
-                    groups: item.groups,
-                    scope: item.scope,
-                    domain,
+                    updatedAt: item.updated_at,
+                    title: item.name,
+                    tags: item.tags,
+                    category: item.category,
+                    revision: item.revision_number,
+                    wip: item.wip,
+                    message: item.message,
+                    teamName,
                     digest: targetFile.getDigestSync(),
                   }
                   return Promise.resolve()
@@ -355,15 +336,17 @@ export default class DocBaseSectionView extends SectionViewBase {
 
             // In case of updating.
             // Call update API.
-            return docbaseService.updateItem(domain, itemId, title, fileBody, tags, groups, docbaseScope)
+            return esaService.updateItem(teamName, itemId, title, fileBody, tags, esaCategory)
               .then((item) => {
                 // Save meta data to config.
-                itemsConfig[item.id].createdAt = item.created_at
-                itemsConfig[item.id].title = item.title
-                itemsConfig[item.id].tags = item.tags.map(tag => tag.name)
-                itemsConfig[item.id].groups = item.groups
-                itemsConfig[item.id].scope = item.scope
-                itemsConfig[item.id].digest = targetFile.getDigestSync()
+                itemsConfig[item.number].updatedAt = item.updated_at
+                itemsConfig[item.number].title = item.name
+                itemsConfig[item.number].tags = item.tags
+                itemsConfig[item.number].category = item.category
+                itemsConfig[item.number].revision = item.revision_number
+                itemsConfig[item.number].wip = item.wip
+                itemsConfig[item.number].message = item.message
+                itemsConfig[item.number].digest = targetFile.getDigestSync()
                 return Promise.resolve()
               })
           })
@@ -372,7 +355,7 @@ export default class DocBaseSectionView extends SectionViewBase {
 
       handledProcess.then(() => {
         // Save meta data to config.
-        config.docbase.items = itemsConfig
+        config.esa.items = itemsConfig
         // See [javascript - sort object properties and JSON.stringify - Stack Overflow](https://stackoverflow.com/questions/16167581/).
         const configSavePromise = configFile.write(stringify(config, { space: 2 }))
         const secretConfigSavePromise = secretConfigFile.write(stringify(secretConfig, { space: 2 }))
@@ -384,7 +367,7 @@ export default class DocBaseSectionView extends SectionViewBase {
       }).catch((error) => {
         // Handle error.
         atom.notifications.addError('Something went wrong.')
-        console.error('Error occurred in DocBaseService: ', error)
+        console.error('Error occurred in EsaService: ', error)
         progressModal.destroy()
       })
     })
