+++ lib/views/section/qiita-section-view.js
@@ -7,33 +7,33 @@ import etch from 'etch'
 import fs from 'fs-plus'
 import stringify from 'json-stable-stringify'
 import SectionViewBase from './section-view-base'
-import EsaService from '../services/esa-service'
-import ConfigUtil from '../utils/config-util'
+import QiitaService from '../../services/qiita-service'
+import ConfigUtil from '../../utils/config-util'
 
-export default class EsaSectionView extends SectionViewBase {
+export default class QiitaSectionView extends SectionViewBase {
   render() {
     return (
-      <section className="esa service">
-        <h2>Sync with esa.io</h2>
+      <section className="qiita service">
+        <h2>Sync with Qiita</h2>
 
-        <label htmlFor="esa-access-token">
+        <label htmlFor="qiita-access-token">
           <span>AccessToken</span>
           <input
             type="text"
-            id="esa-access-token"
-            ref="esaAccessToken"
-            value={this.props.esaAccessToken || ''}
+            id="qiita-access-token"
+            ref="qiitaAccessToken"
+            value={this.props.qiitaAccessToken || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="esa-team-name">
-          <span>TeamName</span>
+        <label htmlFor="qiita-user-name">
+          <span>UserName</span>
           <input
             type="text"
-            id="esa-team-name"
-            ref="esaTeamName"
-            value={this.props.esaTeamName || ''}
+            id="qiita-user-name"
+            ref="qiitaUserName"
+            value={this.props.qiitaUserName || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
@@ -41,84 +41,71 @@ export default class EsaSectionView extends SectionViewBase {
         <h3>
           Item properties&nbsp;
           <a
-            href={this.props.esaItemUrl || ''}
-            style={{ visibility: (this.props.esaItemUrl === null) ? 'hidden' : 'visible' }}
+            href={this.props.qiitaItemUrl || ''}
+            style={{ visibility: (this.props.qiitaItemUrl === null) ? 'hidden' : 'visible' }}
             tabIndex={this.props.startTabIndex++}
           >
             <span role="img" aria-label="Link">🔗</span>
           </a>
         </h3>
 
-        <label htmlFor="esa-title">
+        <label htmlFor="qiita-title">
           <span>Title</span>
           <input
             type="text"
-            id="esa-title"
-            ref="esaTitle"
-            value={this.props.esaTitle || ''}
+            id="qiita-title"
+            ref="qiitaTitle"
+            value={this.props.qiitaTitle || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="esa-tags">
+        <label htmlFor="qiita-tags">
           <span>Tags</span>
           <input
             type="text"
-            id="esa-tags"
-            ref="esaTags"
-            value={this.props.esaTags || ''}
+            id="qiita-tags"
+            ref="qiitaTags"
+            value={this.props.qiitaTags || ''}
             tabIndex={this.props.startTabIndex++}
           />
         </label><br />
 
-        <label htmlFor="esa-category">
-          <span>Category</span>
-          <input
-            type="text"
-            id="esa-category"
-            ref="esaCategory"
-            value={this.props.esaCategory || ''}
-            tabIndex={this.props.startTabIndex++}
-          />
-        </label><br />
-
-        <button id="esa-import" on={{ click: this.importFromEsa }} tabIndex={this.props.startTabIndex++}>
+        <button id="qiita-import" on={{ click: this.importFromQiita }} tabIndex={this.props.startTabIndex++}>
           Import
         </button>
-        <button id="esa-export" on={{ click: this.exportToEsa }} tabIndex={this.props.startTabIndex++}>
+        <button id="qiita-export" on={{ click: this.exportToQiita }} tabIndex={this.props.startTabIndex++}>
           Export
         </button>
       </section>
     )
   }
 
-  importFromEsa(event) {
-    this.handleEsaEvent(event, 'import')
+  importFromQiita(event) {
+    this.handleQiitaEvent(event, 'import')
   }
 
-  exportToEsa(event) {
-    this.handleEsaEvent(event, 'export')
+  exportToQiita(event) {
+    this.handleQiitaEvent(event, 'export')
   }
 
-  handleEsaEvent(event, type = 'import') {
+  handleQiitaEvent(event, type = 'import') {
     try {
-      console.log(`Start Esa ${type}.`)
+      console.log(`Start Qiita ${type}.`)
       const { localPath, configDirPath, optionKeepFilePath } = this.props
-      const accessToken = this.refs.esaAccessToken.value
-      const teamName = this.refs.esaTeamName.value
-      const esaTitle = this.refs.esaTitle.value
-      const esaTags = this.refs.esaTags.value
-      const esaCategory = this.refs.esaCategory.value
-      EsaSectionView.startHandlingEsaEvent(
+      const accessToken = this.refs.qiitaAccessToken.value
+      const userName = this.refs.qiitaUserName.value
+      const qiitaTitle = this.refs.qiitaTitle.value
+      const qiitaTags = this.refs.qiitaTags.value
+      QiitaSectionView.startHandlingQiitaEvent(
         type,
         configDirPath,
         localPath,
         optionKeepFilePath,
         accessToken,
-        teamName,
-        esaTitle,
-        esaTags,
-        esaCategory,
+        userName,
+        qiitaTitle,
+        qiitaTags,
       )
     } catch (error) {
       atom.notifications.addError('Something went wrong.')
@@ -126,19 +113,23 @@ export default class EsaSectionView extends SectionViewBase {
     }
   }
 
-  static startHandlingEsaEvent(
+  static startHandlingQiitaEvent(
     type,
     configDirPath,
     localPath,
     optionKeepFilePath,
     accessToken,
-    teamName,
-    esaTitle,
-    esaTags,
-    esaCategory,
+    userName,
+    qiitaTitle,
+    qiitaTags,
   ) {
     if (localPath === '') {
       atom.notifications.addError('LocalPath must be specified.')
+      return
+    }
+
+    if (type === 'export' && qiitaTags === '') {
+      atom.notifications.addError('Need to specify tag in exporting.')
       return
     }
 
@@ -161,28 +152,28 @@ export default class EsaSectionView extends SectionViewBase {
 
     const loadConfigPromise = configFile.read().then((configBody) => {
       const config = JSON.parse(configBody)
-      config.esa = config.esa || {}
-      config.esa.teamName = teamName
+      config.qiita = config.qiita || {}
+      config.qiita.userName = userName
       // TODO For legacy implementation. To be removed.
-      delete config.esa.accessToken
+      delete config.qiita.accessToken
       return config
     })
     const loadSecretConfigPromise = secretConfigFile.read().then((configBody) => {
       const config = JSON.parse(configBody)
-      config.esa = config.esa || {}
+      config.qiita = config.qiita || {}
       if (atom.config.get('sync-sync.keepSecrets')) {
-        config.esa.accessToken = accessToken
+        config.qiita.accessToken = accessToken
       } else {
-        delete config.esa.accessToken
+        delete config.qiita.accessToken
       }
       return config
     })
 
     Promise.all([loadConfigPromise, loadSecretConfigPromise]).then((configs) => {
       const [config, secretConfig] = configs
-      const itemsConfig = config.esa.items || {}
+      const itemsConfig = config.qiita.items || {}
 
-      const esaService = new EsaService(accessToken)
+      const qiitaService = new QiitaService(accessToken)
       const progressModal = this.showProgressModal('Now Processing', 'Please wait for the process finished...')
       let handledProcess
 
@@ -203,23 +194,18 @@ export default class EsaSectionView extends SectionViewBase {
         //
 
         // Call listing API.
-        handledProcess = esaService.listAllItems(teamName)
+        handledProcess = qiitaService.listAllItems(userName)
           .then((items) => {
             // Save files.
             const fileSaveProcesses = []
             items.forEach((item) => {
-              // Use category as directory.
-              let fileDirectory = ''
-              if (item.category && item.category !== '') {
-                fileDirectory = `${item.category}/`
-              }
               let file = new File(`${localPath}/${item.title}.md`)
               if (optionKeepFilePath && itemsConfig[item.id]) {
                 file = new File(`${configDir.getPath()}/${itemsConfig[item.id].path}`)
               }
               if (file.existsSync()) {
                 // Check not-synced local modification.
-                const syncedDigest = (itemsConfig[item.number]) ? itemsConfig[item.number].digest : null
+                const syncedDigest = (itemsConfig[item.id]) ? itemsConfig[item.id].digest : null
                 if (file.getDigestSync() !== syncedDigest) {
                   atom.notifications.addError(
                     'Cannot import because of not-synced local modification.',
@@ -230,22 +216,18 @@ export default class EsaSectionView extends SectionViewBase {
               }
 
               const relativeItemPath = configDir.relativize(file.getPath())
-              itemsConfig[item.number] = {
+              itemsConfig[item.id] = {
                 path: relativeItemPath,
                 url: item.url,
                 updatedAt: item.updated_at,
-                title: item.name,
-                tags: item.tags,
-                category: item.category,
-                revision: item.revision_number,
-                wip: item.wip,
-                message: item.message,
-                teamName,
+                title: item.title,
+                tags: item.tags.map(tag => tag.name),
+                userName,
               }
 
-              const fileSaveProcess = file.write(this.normalizeNewLine(item.body_md)).then(() => {
+              const fileSaveProcess = file.write(this.normalizeNewLine(item.body)).then(() => {
                 console.log('File saved: ', file.getPath())
-                itemsConfig[item.number].digest = file.getDigestSync()
+                itemsConfig[item.id].digest = file.getDigestSync()
                 return Promise.resolve()
               })
               fileSaveProcesses.push(fileSaveProcess)
@@ -270,11 +252,11 @@ export default class EsaSectionView extends SectionViewBase {
           }
 
           // Call get API.
-          handledProcess = esaService.getItem(teamName, itemId)
+          handledProcess = qiitaService.getItem(itemId)
             .then((item) => {
               if (targetFile.existsSync()) {
                 // Check not-synced local modification.
-                const syncedDigest = (itemsConfig[item.number]) ? itemsConfig[item.number].digest : null
+                const syncedDigest = (itemsConfig[item.id]) ? itemsConfig[item.id].digest : null
                 if (targetFile.getDigestSync() !== syncedDigest) {
                   atom.notifications.addError(
                     'Cannot import because of not-synced local modification.',
@@ -285,16 +267,12 @@ export default class EsaSectionView extends SectionViewBase {
               }
 
               // Save file.
-              return targetFile.write(this.normalizeNewLine(item.body_md)).then(() => {
+              return targetFile.write(this.normalizeNewLine(item.body)).then(() => {
                 console.log('File saved: ', targetFile.getPath())
-                itemsConfig[item.number].updatedAt = item.updated_at
-                itemsConfig[item.number].title = item.name
-                itemsConfig[item.number].tags = item.tags
-                itemsConfig[item.number].category = item.category
-                itemsConfig[item.number].revision = item.revision_number
-                itemsConfig[item.number].wip = item.wip
-                itemsConfig[item.number].message = item.message
-                itemsConfig[item.number].digest = targetFile.getDigestSync()
+                itemsConfig[item.id].updatedAt = item.updated_at
+                itemsConfig[item.id].title = item.title
+                itemsConfig[item.id].tags = item.tags.map(tag => tag.name)
+                itemsConfig[item.id].digest = targetFile.getDigestSync()
                 return Promise.resolve()
               })
             })
@@ -304,30 +282,26 @@ export default class EsaSectionView extends SectionViewBase {
           //
           handledProcess = targetFile.read().then((fileBody) => {
             const relativeItemPath = configDir.relativize(localPath)
-            let title = esaTitle
+            let title = qiitaTitle
             if (!title || title === '') {
               title = targetFile.getBaseName().replace(/\.[^/.]+$/, '') // remove extension
             }
             const itemId = Object.keys(itemsConfig)
               .find(key => itemsConfig[key].path === relativeItemPath)
-            const tags = esaTags.split(',').map(tag => tag.trim())
+            const tags = qiitaTags.split(',').map(tag => tag.trim())
             if (!itemId) {
               // In case of posting.
               // Call post API.
-              return esaService.postItem(teamName, title, fileBody, tags, esaCategory)
+              return qiitaService.postItem(title, fileBody, tags)
                 .then((item) => {
                   // Save meta data to config.
-                  itemsConfig[item.number] = {
+                  itemsConfig[item.id] = {
                     path: relativeItemPath,
                     url: item.url,
                     updatedAt: item.updated_at,
-                    title: item.name,
-                    tags: item.tags,
-                    category: item.category,
-                    revision: item.revision_number,
-                    wip: item.wip,
-                    message: item.message,
-                    teamName,
+                    title: item.title,
+                    tags: item.tags.map(tag => tag.name),
+                    userName,
                     digest: targetFile.getDigestSync(),
                   }
                   return Promise.resolve()
@@ -336,17 +310,13 @@ export default class EsaSectionView extends SectionViewBase {
 
             // In case of updating.
             // Call update API.
-            return esaService.updateItem(teamName, itemId, title, fileBody, tags, esaCategory)
+            return qiitaService.updateItem(itemId, title, fileBody, tags)
               .then((item) => {
                 // Save meta data to config.
-                itemsConfig[item.number].updatedAt = item.updated_at
-                itemsConfig[item.number].title = item.name
-                itemsConfig[item.number].tags = item.tags
-                itemsConfig[item.number].category = item.category
-                itemsConfig[item.number].revision = item.revision_number
-                itemsConfig[item.number].wip = item.wip
-                itemsConfig[item.number].message = item.message
-                itemsConfig[item.number].digest = targetFile.getDigestSync()
+                itemsConfig[item.id].updatedAt = item.updated_at
+                itemsConfig[item.id].title = item.title
+                itemsConfig[item.id].tags = item.tags.map(tag => tag.name)
+                itemsConfig[item.id].digest = targetFile.getDigestSync()
                 return Promise.resolve()
               })
           })
@@ -355,7 +325,7 @@ export default class EsaSectionView extends SectionViewBase {
 
       handledProcess.then(() => {
         // Save meta data to config.
-        config.esa.items = itemsConfig
+        config.qiita.items = itemsConfig
         // See [javascript - sort object properties and JSON.stringify - Stack Overflow](https://stackoverflow.com/questions/16167581/).
         const configSavePromise = configFile.write(stringify(config, { space: 2 }))
         const secretConfigSavePromise = secretConfigFile.write(stringify(secretConfig, { space: 2 }))
@@ -367,7 +337,7 @@ export default class EsaSectionView extends SectionViewBase {
       }).catch((error) => {
         // Handle error.
         atom.notifications.addError('Something went wrong.')
-        console.error('Error occurred in EsaService: ', error)
+        console.error('Error occurred in QiitaService: ', error)
         progressModal.destroy()
       })
     })
