+++ lib/views/section/section-view-base.js
@@ -5,8 +5,8 @@
 import { File } from 'atom'
 import etch from 'etch'
 import fs from 'fs-plus'
-import ViewBase from './view-base'
-import ProgressModalView from './progress-modal-view'
+import ViewBase from '../view-base'
+import ProgressModalView from '../modal/progress-modal-view'
 
 export default class SectionViewBase extends ViewBase {
   static showProgressModal(title, message) {
