+++ lib/views/sync-sync-view.js
@@ -5,9 +5,9 @@
 import { File } from 'atom'
 import etch from 'etch'
 import fs from 'fs-plus'
-import DocBaseSectionView from './docbase-section-view'
-import EsaSectionView from './esa-section-view'
-import QiitaSectionView from './qiita-section-view'
+import DocBaseSectionView from './section/docbase-section-view'
+import EsaSectionView from './section/esa-section-view'
+import QiitaSectionView from './section/qiita-section-view'
 import ViewBase from './view-base'
 
 export default class SyncSyncView extends ViewBase {
